fixed the bug with add to item from vanilla javascript to stimulus, vanilla javascript was giving bugs sometime
